🐞 fix: Server - added automapper

+++ server/src/Harmony.Application/Services/Reflection/ReflectionService.cs
@@ -1,3 +1,4 @@
+using AutoMapper;
 using Harmony.Application.Contracts;
 using Harmony.Application.Contracts.Requests;
 using Harmony.Application.Contracts.Responses;
@@ -9,10 +10,12 @@ namespace Harmony.Application.Services.Reflection;
 public class ReflectionService : IReflectionService
 {
     private readonly IDataContext context;
+    private readonly IMapper mapper;
 
-    public ReflectionService(IDataContext context)
+    public ReflectionService(IDataContext context, IMapper mapper)
     {
         this.context = context;
+        this.mapper = mapper;
     }
 
     public async Task<Response<int>> CreateReflectionAsync(User user, CreateReflectionRequest model)
@@ -67,9 +70,19 @@ public class ReflectionService : IReflectionService
                 StatusCode = 200,
             });
         }
-        var doc = context.Reflections.Where(x => x.Id == id).First();
-        context.Reflections.Remove(doc);
+        var doc = await context.Reflections.Where(x => x.Id == id).FirstOrDefaultAsync();
 
+        if (doc is null)
+        {
+            return await Task.FromResult(new Response<int>
+            {
+                Data = 1,
+                Message = "Reflection not found",
+                StatusCode = 404,
+            });
+        }
+
+        context.Reflections.Remove(doc);
         await context.SaveChangesAsync();
 
         return await Task.FromResult(new Response<int>
@@ -96,17 +109,8 @@ public class ReflectionService : IReflectionService
             .Where(x => x.UserId == userId)
             .Include(x => x.Mood)
             .Include(x => x.Feelings)
-            .Include(x => x.Activities).Select(x => new GetReflectionsResponse
-            {
-                Id = x.Id,
-                Activities = x.Activities.Select(x => x.Name).ToList(),
-                Feelings = x.Feelings.Select(x => x.Name).ToList(),
-                Description = x.Description,
-                Title = x.Title,
-                Mood = x.Mood.Name,
-                UserId = x.UserId,
-                CreatedAt = x.CreatedAt.ToString().Split('T', StringSplitOptions.TrimEntries).First(),
-            })
+            .Include(x => x.Activities)
+            .Select(x => mapper.Map<GetReflectionsResponse>(x))
             .ToList();
 
         return await Task.FromResult(new Response<List<GetReflectionsResponse>>
